Reorganize app files
