+++ safely_report/static/styles.css
