+++ safely_report/survey/templates/survey/run_survey.html
@@ -33,7 +33,7 @@
       {% endif %}
     </div>
     <div class="d-flex justify-content-between">
-      <a href="{{ url_for('back') }}" class="btn btn-secondary">Back</a>
+      <a href="{{ url_for('survey.back') }}" class="btn btn-secondary">Back</a>
       {{ form.submit(class="btn btn-primary") }}
     </div>
   </form>
