+++ safely_report/survey/templates/survey/submit_survey.html
@@ -7,8 +7,8 @@
     <h2 class="mb-4">Congratulations!</h2>
     <div class="mb-4">You have reached survey end. Are you ready to submit?</div>
     <div class="d-flex justify-content-between">
-      <a href="{{ url_for('back') }}" class="btn btn-secondary">Cancel</a>
-      <a href="{{ url_for('submit') }}" class="btn btn-primary">Submit</a>
+      <a href="{{ url_for('survey.back') }}" class="btn btn-secondary">Cancel</a>
+      <a href="{{ url_for('survey.submit') }}" class="btn btn-primary">Submit</a>
     </div>
   </div>
 </div>
