+++ tests/unit/survey_processor/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from survey_processor import SurveyProcessor
+from safely_report.survey.survey_processor import SurveyProcessor
 
 
 @pytest.fixture(scope="function")
@@ -8,8 +8,8 @@ def survey_processor_instance(mocker):
     """
     An instance of `SurveyProcessor` with mock initiation.
     """
-    mocker.patch("survey_processor.read_xlsform")
-    mocker.patch("survey_processor.SurveySession")
+    mocker.patch("safely_report.survey.survey_processor.read_xlsform")
+    mocker.patch("safely_report.survey.survey_processor.SurveySession")
     survey_processor = SurveyProcessor("", {})
 
     return survey_processor
