+++ tests/unit/survey_processor/test_curr_lang.py
@@ -2,7 +2,7 @@ from unittest.mock import PropertyMock
 
 import pytest
 
-from survey_processor import SurveyProcessor
+from safely_report.survey.survey_processor import SurveyProcessor
 
 
 @pytest.mark.parametrize(
