+++ tests/utils.py
@@ -1,4 +1,4 @@
-from survey_processor import SurveyProcessor
+from safely_report.survey.survey_processor import SurveyProcessor
 
 
 class MockFlaskSession(dict):
